Update app show pages for devices list and individual device
If the user navigates to the devices page a list of the devices that are available
in the database will be shown, in this page if the user selects a specific device
then another page will be shown with only information about that device. In this
page the user has a button to update the information about this device.

Signed-off-by: Nelson Vieira <2080511@student.uma.pt>